bonus

+++ worker/src/DepLibUring.cpp
@@ -357,8 +357,8 @@ DepLibUring::LibUring::LibUring()
 				MS_WARN_TAG(
 				  info,
 				  "io_uring_register_buffers() failed due to low RLIMIT_MEMLOCK (soft:%llu, hard:%llu), disabling zero copy: %s",
-				  l.rlim_cur,
-				  l.rlim_max,
+				  static_cast<unsigned long long>(l.rlim_cur),
+				  static_cast<unsigned long long>(l.rlim_max),
 				  std::strerror(error));
 			}
 		}
